refactor(front: bean card): コンポーネントのディレクトリ移動
