+++ front/src/features/beans/components/molecules/BeanImagesSwiper.tsx
