+++ front/src/features/beans/components/molecules/BeanTasteChart.tsx
